Complete VM variable semantics audit
Test raw module paths from ShString and SACT and make their outputs
explicitly cross the VM word boundary. Fix the remaining SACT word
buffer pointer type and cover the full word range in US/UG stack
operations.

The audit now passes with both int and uint16_t storage typedefs, plus
a long long canary build for pointer propagation.

+++ modules/SACT/sactchart.c
@@ -29,7 +29,8 @@ void schart_pos(vmvar_t *pos, int pos1, int pos2, int val1, int val2, int val) {
 	if (val1 == val2) {
 		*pos = 0;
 	} else {
-		*pos = (((pos2 - pos1) * (val - val1)) / (val2 - val1)) + pos1;
+		*pos = (uint16_t)((((pos2 - pos1) * (val - val1)) /
+				   (val2 - val1)) + pos1);
 	}
 }
